+++ tests/unit_tests/transformer/moe/test_allgather_token_dispatcher.py
@@ -1,0 +1,60 @@
+# Copyright (c) 2023, NVIDIA CORPORATION. All rights reserved.
+
+import pytest
+import torch
+
+from megatron.core.utils import is_te_min_version
+from tests.unit_tests.test_utilities import Utils
+from tests.unit_tests.transformer.moe.test_token_dispatcher import (
+    MoEModelTestContainer,
+    permute_fusion_params,
+)
+
+
+class TestAllgatherDispatcher:
+    def setup_method(self, method):
+        pass
+
+    def teardown_method(self, method):
+        Utils.destroy_model_parallel()
+
+    @pytest.mark.skipif(not torch.cuda.is_available(), reason="CUDA not available")
+    @pytest.mark.internal
+    @pytest.mark.parametrize("tp_size,ep_size", [(8, 1), (1, 8), (2, 4), (1, 1)])
+    @pytest.mark.parametrize("permute_fusion", permute_fusion_params)
+    def test_forward_backward(self, tp_size, ep_size, permute_fusion):
+        container = MoEModelTestContainer(
+            tp_size=tp_size,
+            ep_size=ep_size,
+            pp_size=1,
+            num_moe_experts=8,
+            moe_router_topk=2,
+            moe_router_load_balancing_type="aux_loss",
+            moe_token_dispatcher_type="allgather",
+            moe_permute_fusion=permute_fusion,
+        )
+
+        container.dispatcher_dropless_test()
+
+    @pytest.mark.skipif(not torch.cuda.is_available(), reason="CUDA not available")
+    @pytest.mark.internal
+    @pytest.mark.parametrize("permute_fusion", permute_fusion_params)
+    @pytest.mark.parametrize(
+        "tp_size,ep_size,moe_tp_size", [(1, 1, 8), (1, 2, 4), (1, 4, 2), (2, 2, 4)]
+    )
+    def test_moe_tp_forward_backward(self, tp_size, ep_size, moe_tp_size, permute_fusion):
+        container = MoEModelTestContainer(
+            tp_size=tp_size,
+            ep_size=ep_size,
+            pp_size=1,
+            moe_tp_size=moe_tp_size,
+            num_moe_experts=8,
+            moe_router_topk=2,
+            moe_router_load_balancing_type="aux_loss",
+            moe_token_dispatcher_type="allgather",
+            sequence_parallel=True,
+            moe_permute_fusion=permute_fusion,
+            use_cpu_initialization=False,
+        )
+
+        container.dispatcher_dropless_test()
